Correct handling of cleaned hgvs codes
errorfixer has been implemented to enable providing cleaned versions of
unparseable hgvs codes. This works by checking in the hgvsparser init,
whether the genomic entry already exists in the errorfixer. If so, the
provided cleaned strings from the hgvs_errors.json file are used.

The initial implementation had not been tested with available cleaned
hgvs strings, thus the implementation did not actually correctly
override the entries, since the returned values was a list of cleaned
hgvs strings, but a genomic_entry objects had been expected instead.

The changes moved the error fixing portion after getting the basic
information from the genomic_entry. The cleaned strings will be
converted to hgvs objects and act as a placein for the variants
attribute.