Switching Mutalyzer to SOAP based API calls
Batch-processing of HGVS codes for reference transcript resolution
requires the usage of the batch system. Because Mutalyzer provides a
SOAP-API with WSDL specification, Zeep was used to automatically
generate methods in the Mutalyzer class.

A system for transcript resolution has been written to work on an entire
list of cases. This will automatically submit a batch process, track its
progress and try to fix incomplete reference sequences.