Current script for generation of old json files
This file contains all necessary steps for the current generation of old
json format files. This file can and should be adapted to ones specific
needs. Any complex functionality, should if possible be put into the
library itself and not in these scripts.

All steps from reading the input directory are performed, some
checkpoints are saved as pickles to recover in case of remote api
failures (frequency might depend on ones internet connection). Print
statements are mainly used as graphical indicators of progress.