Removed VCF generation from case object initialization
Since we need to reconstruct the VCF after correcting the reference
transcripts, I removed the VCf construction from the case object
initialization. The function dump_vcf now creates a sorted VCF if none has yet
been created or it needs to recreated and then dumps it to a given path.
The VCF is saved as an attribute in case it needs to be reused
afterwards. I also changed the original vcf attribute to realvcf. Errors
in the VCF generation are currently just printed.