Rebased the repository and added minor changes to VCF generation
Previously errors in the VCF generation were simply printed. Now the
case ID and the error message are logged and the VCFs are compressed
with bgzip. The location of the Jannovar client currently has to be changed manually. Also, to execute the VCF generation a location where the VCFs are supposed to be saved (e.g. "data/PEDIA/mutations/") has to be given.