Change omim ID assignment in syndrome table
Previously omim ids from syndromes would be coerced to list and
processed. This did not take empty omim id lists into account and these
syndromes would disappear in the subsequent geneList generation. To
ensure the visibility of a selected diagnosis, syndromes without omim id
in the selected syndromes will be assigned a default omim id of 0.