Preprocess adaptation for vcf generation
Sequence of VCF generation and QC have been reworked to properly account
for qc changes because of VCF generation.

Initial qc check will be performed before VCF generation, as to prevent
the generation of VCF files for cases not passing QC.

QC will thereafter perform an additional check for the generated VCF and
record potential errors in the quality_check.log.