Fix incorrect gene name #41
In this commit, we nned to use HGVS dict with version 3.
We add 'correct_gene' in HGVS dict file. Then we load it
in error_fixer. However, some cases have incorrect gene name
but with correct HGVS. Then the clean hgvs array is empty.
Therefore, we check if the array is empty before we choose
if we want to apply the HGVS code from HGVS dict file.