Add preprocessing script for merging DGFM
DGFM_gestalt_syndrome file is for generating gestalt support omim
mapping. However, there are some errors. We need to fixed it by the
mapping from AWS. We then used compare_syn to compare the omim
mapping. In the end, use merge_pedia_dgfm to merge FM and gestalt
with the information from AWS